Fix redirect for declined invitation.
When the user tried to decline an invitation being logged out, he was
requested to log in. The current URL was saved to redirect the user to
this page after he logs in. The issues was the URL being for the "Accept
Invitation" page instead of the "Decline Invitation".

Related to #259